Fixed track removal for amounts over 100 tracks. Got rid of track preview player as am not sure non premium Spotify members can use it and it is causing bugs atm, may implement again later
(cherry picked from commit 70989d8)

+++ app/src/main/java/com/reissgrvs/spotifyplaylisttool/PlaylistUpdateUtils/PlaylistUpdateUtils.java
@@ -2,11 +2,13 @@ package com.reissgrvs.spotifyplaylisttool.PlaylistUpdateUtils;
 
 
 import android.os.AsyncTask;
+import android.util.Log;
 
 import com.reissgrvs.spotifyplaylisttool.CustomTrack;
 import com.reissgrvs.spotifyplaylisttool.SpotifyAPI.SpotifyAPIManager;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -19,12 +21,14 @@ import kaaes.spotify.webapi.android.models.PlaylistTrack;
 import kaaes.spotify.webapi.android.models.Track;
 import kaaes.spotify.webapi.android.models.TrackToRemove;
 import kaaes.spotify.webapi.android.models.TracksToRemove;
+import retrofit.RetrofitError;
+
+import static android.os.Debug.waitForDebugger;
 
 
 public class PlaylistUpdateUtils {
 
     static void updateSpotifyPlaylist(String userID, String playlistID, ArrayList<PlaylistTrack> oldTrackArray, ArrayList<PlaylistTrack> newTrackArray){
-
         //Populate oldTrackSet and newTrackSet
         Set<CustomTrack> oldTrackSet = new HashSet<>();
         for (PlaylistTrack pTrack : oldTrackArray){
@@ -48,7 +52,6 @@ public class PlaylistUpdateUtils {
             }
         }
 
-
         removeTracks(userID, playlistID, removeTracks);
         addTracks(userID, playlistID, newTrackSet);
 
@@ -58,24 +61,51 @@ public class PlaylistUpdateUtils {
     public static void addTracks(String userID, String playlistID, Collection<CustomTrack> newTracks){
         if (!newTracks.isEmpty()) {
 
-            Map<String, Object> body = new HashMap<>();
-            Map<String, Object> query = new HashMap<>();
+            ArrayList<CustomTrack> tracks = new ArrayList<>(newTracks);
+            int pageLength = 50;
 
-            String trackList = "";
-            for (CustomTrack track : newTracks) {
-                trackList += track.mTrack.uri + ",";
+            for(int i = 0; i < (tracks.size() / pageLength); i++){
+                addTrackChunk(userID,playlistID, tracks.subList(pageLength*i, pageLength*(i+1)));
             }
-            query.put("uris", trackList.substring(0, trackList.length() - 1));
 
-            SpotifyAPIManager.getService().addTracksToPlaylist(userID, playlistID, query, body);
+            if(tracks.size() % pageLength > 0){
+                addTrackChunk(userID, playlistID ,tracks.subList(tracks.size() - tracks.size() % pageLength , tracks.size()));
+            }
         }
     }
 
-    private static void removeTracks(String userID, String playlistID, Collection<CustomTrack> oldTracks){
+    private static void addTrackChunk(String userID, String playlistID, Collection<CustomTrack> tracksChunk){
+        List<String> trackList = new ArrayList<>();
+        for (CustomTrack track : tracksChunk){
+            if(track.mTrack.uri.startsWith("spotify:track:")){
+                trackList.add(track.mTrack.uri);
+            }
+        }
 
+        Map<String, Object> body = new HashMap<>();
+        body.put("uris", trackList);
+        try {
+            SpotifyAPIManager.getService().addTracksToPlaylist(userID, playlistID, null, body);
+        }catch (RetrofitError error)
+        {
+            Log.e("retrofiterror" ,error.getResponse().getReason());
+        }
+    }
+    private static void removeTracks(String userID, String playlistID, Collection<CustomTrack> oldTracks){
         if(!oldTracks.isEmpty()) {
-            TracksToRemove tracksToRemove = packageTracksToRemove(oldTracks);
-            SpotifyAPIManager.getService().removeTracksFromPlaylist(userID, playlistID, tracksToRemove);
+            ArrayList<CustomTrack> tracks = new ArrayList<>(oldTracks);
+            int pageLength = 50;
+
+            for(int i = 0; i < (tracks.size() / pageLength); i++){
+                TracksToRemove tracksToRemove = packageTracksToRemove(tracks.subList(i*pageLength, (i+1)*pageLength));
+                SpotifyAPIManager.getService().removeTracksFromPlaylist(userID, playlistID, tracksToRemove);
+            }
+
+            if(tracks.size() % pageLength > 0){
+                TracksToRemove tracksToRemove = packageTracksToRemove(tracks.subList(tracks.size() - tracks.size() % pageLength , tracks.size()));
+                SpotifyAPIManager.getService().removeTracksFromPlaylist(userID, playlistID, tracksToRemove);
+            }
+
         }
     }
